feat:부모 최애 선물 조회

+++ src/main/java/unid/team8/repository/GiftRepository.java
@@ -3,6 +3,6 @@ package unid.team8.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import unid.team8.entity.Gift;
 
-public interface GiftRepository extends JpaRepository<Gift, Integer> {
+public interface GiftRepository extends JpaRepository<Gift, Long> {
     // 기본적인 CRUD 메서드를 JpaRepository가 제공함
 }
